Combat log styling, update docs and plan
